Bug fix: prevent unwanted clicks using menu
This is a strange way to fix the problem
and I am clearly working against RenPy with this.

The hard pauses may be able to be removed.
This may interfere with future menu UI.
For now this seems to work.
When the player has to move, they cannot click,
skip, or afm through the script.